style(table-check-modal): move directory from `others` to `modals/advanced`
