+++ src/feedbacks/modals/advanced/table-modal/PTableCheckModal.vue
